+++ app/src/main/java/jp/ac/it_college/std/ikemen/reachable/company/coupon/CouponSelectFragment.java
@@ -41,7 +41,7 @@ import jp.ac.it_college.std.ikemen.reachable.company.view.EmptySupportRecyclerVi
 import jp.ac.it_college.std.ikemen.reachable.company.MainActivity;
 import jp.ac.it_college.std.ikemen.reachable.company.OnActionClickListener;
 import jp.ac.it_college.std.ikemen.reachable.company.R;
-import jp.ac.it_college.std.ikemen.reachable.company.RecyclerItemClickListener;
+import jp.ac.it_college.std.ikemen.reachable.company.view.listener.RecyclerItemClickListener;
 import jp.ac.it_college.std.ikemen.reachable.company.aws.UploadObservers;
 import jp.ac.it_college.std.ikemen.reachable.company.aws.AwsUtil;
 import jp.ac.it_college.std.ikemen.reachable.company.aws.S3UploadManager;
